refs #63 Move interface to common directory
