+++ common/IAuth.ts
