+++ common/IBook.ts
@@ -1,4 +1,5 @@
 import { Timestamp } from './Timestamp';
+
 export default interface IBook {
   Title: string;
   ISBN: string;
@@ -8,15 +9,16 @@ export default interface IBook {
   Publisher: string;
   Comment: string;
   Created: Date;
-  CreatedUserId: string;
-  CreatedUserName: string;
+  CreatedUserId: string | null;
+  CreatedInfo: string;
   Modified: Date;
-  ModifiedUserId: string;
-  ModifiedUserName: string;
+  ModifiedUserId: string | null;
+  ModifiedInfo: string;
   Location: string;
-  OnLoan?: boolean;
-  LastBorrowUserId?: string;
-  LastBorrowTimestamp?: null | Timestamp;
+  OnLoan: boolean | null;
+  LastBorrowUserId: string | null;
+  // TODO: Union type に firebase.firestore.FieldValue 追加する
+  LastBorrowTimestamp: Timestamp | null;
   Save(): any;
   Rent(userId: string): Promise<void>;
   Return(): Promise<void>;
