+++ functions/src/onWriteBook.ts
@@ -1,9 +1,11 @@
 import * as admin from 'firebase-admin';
 import * as functions from 'firebase-functions';
 
-// FIXME: firebase deploy 時に @common/ が解釈できないため、相対パス指定に戻している
+// FIXME: firebase deploy 時に @common/booksManagementEvent が解釈できないため、相対パス指定に戻している
+// src/model/Audit.ts などの定義と何が違う？
 import * as booksManagementEvent from '../../common/booksManagementEvent';
 import * as moment from 'moment';
+import IBook from '@common/IBook';
 
 const db = admin.firestore();
