+++ src/views/BooksList.vue
@@ -44,9 +44,9 @@
 import { Component, Vue } from "vue-property-decorator";
 import { BooksModule } from "@/modules/BooksModule";
 import { SignInModule } from "@/modules/SignInModule";
-import IBook from "@/model/IBook";
+import IBook from "@common/IBook";
 import Book from "@/model/Book";
-import BookDetail from "../components/BookDetail.vue";
+import BookDetail from "@/components/BookDetail.vue";
 
 const Super = Vue.extend({
   methods: BooksModule.mapActions(["updateList", "setCurrentBook"])
